third_party/WebKit/LayoutTests/bluetooth/characteristic/notifications/concurrent-starts.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script src="../../../resources/bluetooth/web-bluetooth-test.js"></script>
<script src="../../../resources/mojo-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  return getMeasurementIntervalCharacteristic()
    .then(({characteristic, fake_characteristic}) => {
      return fake_characteristic
        .setNextSubscribeToNotificationsResponse(
          GATT_SUCCESS)
        .then(() => Promise.all([
          characteristic.startNotifications(),
          characteristic.startNotifications(),
          characteristic.startNotifications()]))
        .then(() =>
          fake_characteristic.setNextUnsubscribeFromNotificationsResponse(
            GATT_SUCCESS))
        .then(() => fake_characteristic.isNotifying())
        .then(is_notifying => assert_true(is_notifying))
      });
}, 'Multiple starts in a row.');
</script>
